Add counters to Quart
